refactor files and functions
